model::operator>>: now it can read floating-point coordinates

+++ model.cpp
@@ -72,19 +72,19 @@ std::istream &operator>>(std::istream &in, Model &model) {
             throw model_format_error("unable to read fngure type");
         }
         if (type == "segment") {
-            int x1, y1, x2, y2;
+            double x1, y1, x2, y2;
             if (!(in >> x1 >> y1 >> x2 >> y2)) {
                 throw model_format_error("unable to read segment");
             }
             model.addFigure(std::make_shared<figures::Segment>(Point(x1, y1), Point(x2, y2)));
         } else if (type == "rectangle") {
-            int x1, y1, x2, y2;
+            double x1, y1, x2, y2;
             if (!(in >> x1 >> y1 >> x2 >> y2)) {
                 throw model_format_error("unable to read rectangle");
             }
             model.addFigure(std::make_shared<figures::Rectangle>(BoundingBox({Point(x1, y1), Point(x2, y2)})));
         } else if (type == "ellipse") {
-            int x1, y1, x2, y2;
+            double x1, y1, x2, y2;
             if (!(in >> x1 >> y1 >> x2 >> y2)) {
                 throw model_format_error("unable to read ellipse");
             }
